feat(tree): write level_order with recursion
The key insight was the necessary &block to pass
the block to the next recursive call. Or else the
method loses track of the block that was given.

Another key insight was the need to handle the
block_given? condition where we want to return
nil and print as a side effect. The final line
handles this correctly to return nil instead of
the array of data values.